cc-3310 add float test case

+++ Bundles/ProductMeasurementUnit/src/Spryker/Zed/ProductMeasurementUnit/Business/Model/ProductMeasurementSalesUnit/ProductMeasurementSalesUnitValue.php
@@ -44,13 +44,13 @@ class ProductMeasurementSalesUnitValue implements ProductMeasurementSalesUnitVal
      * and 2 KG represents 1 quantity (unit to availability conversion ratio is 0.5).
      * The retrieved normalized unit value is 1600 (16.00 KG when displayed).
      *
-     * @param int $availabilityValue
+     * @param float $availabilityValue
      * @param float $unitToAvailabilityConversion
      * @param int $unitPrecision
      *
      * @return int
      */
-    protected function calculateNormalizedValue(int $availabilityValue, float $unitToAvailabilityConversion, int $unitPrecision): int
+    protected function calculateNormalizedValue(float $availabilityValue, float $unitToAvailabilityConversion, int $unitPrecision): int
     {
         return (int)round(
             $this->calculateFloatNormalizedValue($availabilityValue, $unitToAvailabilityConversion, $unitPrecision)
@@ -60,13 +60,13 @@ class ProductMeasurementSalesUnitValue implements ProductMeasurementSalesUnitVal
     /**
      * @see ProductMeasurementSalesUnitValue::calculateNormalizedValue()
      *
-     * @param int $availabilityValue
+     * @param float $availabilityValue
      * @param float $unitToAvailabilityConversion
      * @param int $unitPrecision
      *
      * @return float
      */
-    protected function calculateFloatNormalizedValue(int $availabilityValue, float $unitToAvailabilityConversion, int $unitPrecision): float
+    protected function calculateFloatNormalizedValue(float $availabilityValue, float $unitToAvailabilityConversion, int $unitPrecision): float
     {
         return $availabilityValue / $unitToAvailabilityConversion * $unitPrecision;
     }
